+++ CHANGELOG.md
@@ -6,3 +6,4 @@ Put changelog here:
 
 -----------------
 - Fixed structure upgrading not saving anything.
+- Added `SpecialModelRenderersEntrypoint` and `SpecialModelRenderersEvents` for ease of use when creating new special item renderers for use with Block Entities.